TypeChecker: if now typechecked as well as binary operator bugfix
Binary operators did not resolve their return type symbol, and instead
just returned the symbol, which is not how the current version of the
typechecker operators (we operator on raw type values, not symbols).
In the future it would be more efficient, easier, and less squirrelly
to just compare type symbols for typechecking, but who knows.

'if' expressions now have both of their branches checked to ensure
that they return the proper type, and the same type, so that the
if expression doesn't allow for dynamic typing, basically.