Division and Bitshifting operators
We put this off for a while because these instructions are more
complicated in 'x86_64' assembly, and I am lazy!

We require more work done on the lexer before we can properly lex and
parse binary operators that are longer than one character, so for now
bitshifting is designated with square brackets... It's temporary!