¡Arrays!
We have arrays working in Aly. This took a very long time and I am very happy that we managed to get this implemented.Whilst this works, we are being very strict with how arrays work right now. We mainly target integers only but that is by design for testing!